From Wim Lavrijsen: macosx changes attached. It fixes the warning that Eddy observed, as well as an unheeded warning that comes from the python headers if you install a "non-framework" python.
git-svn-id: http://root.cern.ch/svn/root/trunk@10494 27541ba8-7e3a-0410-8455-c3a389f83636

+++ pyroot/src/MethodDispatcher.cxx
@@ -1,4 +1,4 @@
-// @(#)root/pyroot:$Name:  $:$Id: MethodDispatcher.cxx,v 1.7 2004/08/11 04:42:11 brun Exp $
+// @(#)root/pyroot:$Name:  $:$Id: MethodDispatcher.cxx,v 1.8 2004/11/05 09:05:45 brun Exp $
 // Author: Wim Lavrijsen, Apr 2004
 
 // Bindings
@@ -92,7 +92,7 @@ PyObject* PyROOT::MethodDispatcher::operator()( PyObject* aTuple, PyObject* aDic
    }
 
    char txt[ 256 ];
-   sprintf( txt, "none of the %d overloaded methods succeeded", m_methods.size() );
+   sprintf( txt, "none of the %d overloaded methods succeeded", (int) m_methods.size() );
    PyErr_SetString( PyExc_TypeError, txt );
    return 0;
 }
